% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEcR_style.R
\name{WEcR_style}
\alias{WEcR_style}
\title{Add bbc theme to ggplot chart}
\usage{
WEcR_style()
}
\description{
This function allows you to add the WEcR theme to your ggplotgraphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
bbc_style()
}
\keyword{bbc_style}
